test adjustments and fixes
